deinterlace: t7/t3 config of afbce addr [1/1]
PD#SWPL-54771

Problem:
afbce mode heard addr not corect
di dts config update

Solution:
modify the afbce heard addr
modify the t7 dts config

Verify:
T7

+++ drivers/amlogic/media/deinterlace/di_pqa.h
@@ -413,8 +413,8 @@ struct afbcd_ctr_s {
 	unsigned int l_h;
 	unsigned int l_w;
 	unsigned int l_bitdepth;
-	unsigned int addr_h;
-	unsigned int addr_b;
+	ulong addr_h;
+	ulong addr_b;
 //	unsigned int mem_addr_h;
 //	unsigned int mem_addr_b;
 	unsigned int pst_in_vtp;
